requirements.txt
psycopg2-binary 
python-dotenv
pandas 
vanna[chromadb,openai,postgres] 
streamlit 
matplotlib 
seaborn
psycopg2
flask_cors